wine-clustering-app/requirements.txt
streamlit
scikit-learn
numpy
pillow
tensorflow